Separate plugin API from implementation
Making it easier to eventually move the API to a separate module altogether.
