+++ src/main/java/org/dependencytrack/plugin/api/ExtensionFactory.java
@@ -16,7 +16,7 @@
  * SPDX-License-Identifier: Apache-2.0
  * Copyright (c) OWASP Foundation. All Rights Reserved.
  */
-package org.dependencytrack.plugin;
+package org.dependencytrack.plugin.api;
 
 import java.io.Closeable;
