+++ src/test/java/org/dependencytrack/plugin/PluginManagerTest.java
@@ -20,6 +20,9 @@ package org.dependencytrack.plugin;
 
 import alpine.model.IConfigProperty;
 import org.dependencytrack.PersistenceCapableTest;
+import org.dependencytrack.plugin.api.ExtensionFactory;
+import org.dependencytrack.plugin.api.ExtensionPoint;
+import org.dependencytrack.plugin.api.Plugin;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.contrib.java.lang.system.EnvironmentVariables;
@@ -42,7 +45,7 @@ public class PluginManagerTest extends PersistenceCapableTest {
     @Test
     public void testGetLoadedPlugins() {
         final List<Plugin> loadedPlugins = PluginManager.getInstance().getLoadedPlugins();
-        assertThat(loadedPlugins).satisfiesExactly(plugin -> assertThat(plugin.name()).isEqualTo("dummy123"));
+        assertThat(loadedPlugins).satisfiesExactly(plugin -> assertThat(plugin.name()).isEqualTo("dummy"));
         assertThat(loadedPlugins).isUnmodifiable();
     }
