+++ src/test/resources/META-INF/services/org.dependencytrack.plugin.api.ExtensionPointMetadata
